Build 64-bit variants of emulator, more RV64G improvements

+++ lib/libriscv/rv64i.hpp
@@ -17,7 +17,7 @@ namespace riscv
 
 		static std::string to_string(CPU<8>& cpu, format_t format, const instruction_t& instr);
 
-		static inline uint64_t SRA(bool is_signed, uint64_t shifts, uint64_t value)
+		static inline uint64_t SRA(bool is_signed, uint32_t shifts, uint64_t value)
 		{
 			const uint64_t sign_bits = -is_signed ^ 0x0;
 			const uint64_t sign_shifted = sign_bits << (64 - shifts);
